increase FBX importer object ID storage size
Some exporters write object IDs beyond 2^32

+++ cgogn/io/fbx/fbx_importer.h
@@ -69,7 +69,7 @@ public:
 
 protected:
 	using Skeleton = AnimationSkeleton;
-	using ObjectId = uint32;
+	using ObjectId = uint64;
 	using AnimTimeT = double;
 	using AnimScalar = geometry::Scalar;
 	using AnimationT = geometry::KeyframedAnimation<std::vector, AnimTimeT, AnimScalar>;
